Alteração em layoutxml do mdf-e para aceitar mdf-e 3.0
