+++ MDFe.Classes/Extencoes/ExtMDFeConsStatServMDFe.cs
@@ -30,13 +30,46 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-using System.Xml.Serialization;
 
-namespace MDFe.Classes.Servicos.Flags
+using System.Xml;
+using DFe.Utils;
+using MDFe.Classes.Informacoes.StatusServico;
+using MDFe.Utils.Configuracoes;
+using MDFe.Utils.Validacao;
+
+namespace MDFe.Classes.Extencoes
 {
-    public enum VersaoServico
+    public static class ExtMDFeConsStatServMDFe
     {
-        [XmlEnum("1.00")]
-        Versao100 = 100
+        public static void ValidarSchema(this MDFeConsStatServMDFe consStatServMDFe)
+        {
+            var xmlValidacao = consStatServMDFe.XmlString();
+
+            Validador.Valida(xmlValidacao, "consStatServMDFe_v1.00.xsd");
+        }
+
+        public static string XmlString(this MDFeConsStatServMDFe consStatServMDFe)
+        {
+            return FuncoesXml.ClasseParaXmlString(consStatServMDFe);
+        }
+
+        public static XmlDocument CriaRequestWs(this MDFeConsStatServMDFe consStatServMdFe)
+        {
+            var request = new XmlDocument();
+            request.LoadXml(consStatServMdFe.XmlString());
+
+            return request;
+        }
+
+        public static void SalvarXmlEmDisco(this MDFeConsStatServMDFe consStatServMdFe)
+        {
+            if (MDFeConfiguracao.NaoSalvarXml()) return;
+
+            var caminhoXml = MDFeConfiguracao.CaminhoSalvarXml;
+
+            var arquivoSalvar = caminhoXml + @"\-pedido-status-servico.xml";
+
+            FuncoesXml.ClasseParaArquivoXml(consStatServMdFe, arquivoSalvar);
+        }
     }
 }
