+++ MDFe.Classes/Extencoes/ExtMDFeEvEncMDFe.cs
@@ -30,78 +30,25 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-using System.Xml;
+
 using DFe.Utils;
-using DFe.Utils.Assinatura;
-using MDFe.Classes.Informacoes.Evento;
 using MDFe.Classes.Informacoes.Evento.CorpoEvento;
-using MDFe.Utils.Configuracoes;
 using MDFe.Utils.Validacao;
 
-namespace MDFe.Utils.Extencoes
+namespace MDFe.Classes.Extencoes
 {
-    public static class ExtMDFeEventoMDFe
+    public static class ExtMDFeEvEncMDFe
     {
-        public static void ValidarSchema(this MDFeEventoMDFe evento)
+        public static void ValidaSchema(this MDFeEvEncMDFe evEncMDFe)
         {
-            var xmlValido = evento.XmlString();
+            var xmlEncerramento = evEncMDFe.XmlString();
 
-            Validador.Valida(xmlValido, "eventoMDFe_v1.00.xsd");
-
-            var tipoEvento = evento.InfEvento.DetEvento.EventoContainer.GetType();
-
-            if (tipoEvento == typeof (MDFeEvCancMDFe))
-            {
-                var objetoXml = (MDFeEvCancMDFe) evento.InfEvento.DetEvento.EventoContainer;
-                objetoXml.ValidaSchema();
-            }
-
-            if (tipoEvento == typeof (MDFeEvEncMDFe))
-            {
-                var objetoXml = (MDFeEvEncMDFe)evento.InfEvento.DetEvento.EventoContainer;
-
-                objetoXml.ValidaSchema();
-            }
-
-            if (tipoEvento == typeof (MDFeEvIncCondutorMDFe))
-            {
-                var objetoXml = (MDFeEvIncCondutorMDFe)evento.InfEvento.DetEvento.EventoContainer;
-
-                objetoXml.ValidaSchema();
-            }
-
-
+            Validador.Valida(xmlEncerramento, "evEncMDFe_v1.00.xsd");
         }
 
-        public static XmlDocument CriaXmlRequestWs(this MDFeEventoMDFe evento)
+        public static string XmlString(this MDFeEvEncMDFe evEncMDFe)
         {
-            var xmlRequest = new XmlDocument();
-            xmlRequest.LoadXml(evento.XmlString());
-
-            return xmlRequest;
+            return FuncoesXml.ClasseParaXmlString(evEncMDFe);
         }
-
-        public static string XmlString(this MDFeEventoMDFe evento)
-        {
-            return FuncoesXml.ClasseParaXmlString(evento);
-        }
-
-        public static void Assinar(this MDFeEventoMDFe evento)
-        {
-            evento.Signature = AssinaturaDigital.Assina(evento, evento.InfEvento.Id,
-                MDFeConfiguracao.X509Certificate2);
-        }
-
-        public static void SalvarXmlEmDisco(this MDFeEventoMDFe evento, string chave)
-        {
-            if (MDFeConfiguracao.NaoSalvarXml()) return;
-
-            var caminhoXml = MDFeConfiguracao.CaminhoSalvarXml;
-
-            var arquivoSalvar = caminhoXml + @"\" + chave + "-ped-eve.xml";
-
-            FuncoesXml.ClasseParaArquivoXml(evento, arquivoSalvar);
-        }
-
     }
 }
