+++ MDFe.Classes/Flags/MDFeIdentificacaoProcessoEmissao.cs
@@ -39,6 +39,6 @@ namespace MDFe.Classes.Flags
         [XmlEnum("0")]
         EmissaoComAplicativoContribuinte = 0,
         [XmlEnum("1")]
-        EmissoPeloContribuinteComAplicativoFornecidoPeloFisco = 1
+        EmissoPeloContribuinteComAplicativoFornecidoPeloFisco = 3
     }
 }
