+++ MDFe.Classes/Flags/MDFeTpUnidTransp.cs
@@ -45,7 +45,7 @@ namespace MDFe.Classes.Flags
         [XmlEnum("4")]
         Balsa = 4,
         [XmlEnum("5")]
-        Aeronavel = 5,
+        Aeronave = 5,
         [XmlEnum("6")]
         Vagao = 6,
         [XmlEnum("7")]
