+++ MDFe.Classes/Informacoes/ConsultaNaoEncerrados/MDFeCosMDFeNaoEnc.cs
@@ -33,7 +33,7 @@
 using System;
 using System.Xml.Serialization;
 using DFe.Classes.Flags;
-using MDFe.Classes.Servicos.Flags;
+using MDFe.Utils.Flags;
 
 namespace MDFe.Classes.Informacoes.ConsultaNaoEncerrados
 {
