+++ MDFe.Classes/Retorno/MDFeConsultaNaoEncerrado/MDFeRetConsMDFeNao.cs
@@ -38,7 +38,7 @@ using DFe.Classes.Extencoes;
 using DFe.Classes.Flags;
 using DFe.Utils;
 using MDFe.Classes.Informacoes.ConsultaNaoEncerrados;
-using MDFe.Classes.Servicos.Flags;
+using MDFe.Utils.Flags;
 
 namespace MDFe.Classes.Retorno.MDFeConsultaNaoEncerrado
 {
