+++ MDFe.Utils/Flags/VersaoServico.cs
@@ -30,45 +30,17 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-using System.Xml;
-using DFe.Utils;
-using MDFe.Classes.Informacoes.ConsultaProtocolo;
-using MDFe.Utils.Configuracoes;
-using MDFe.Utils.Validacao;
 
-namespace MDFe.Utils.Extencoes
+using System.Xml.Serialization;
+
+namespace MDFe.Utils.Flags
 {
-    public static class ExtMDFeConsSitMDFe
+    public enum VersaoServico
     {
-        public static void ValidarSchema(this MDFeConsSitMDFe consSitMdfe)
-        {
-            var xmlEnvio = consSitMdfe.XmlString();
+        [XmlEnum("1.00")]
+        Versao100 = 100,
 
-            Validador.Valida(xmlEnvio, "consSitMdfe_v1.00.xsd");
-        }
-
-        public static string XmlString(this MDFeConsSitMDFe consSitMdfe)
-        {
-            return FuncoesXml.ClasseParaXmlString(consSitMdfe);
-        }
-
-        public static XmlDocument CriaRequestWs(this MDFeConsSitMDFe consSitMdfe)
-        {
-            var request = new XmlDocument();
-            request.LoadXml(consSitMdfe.XmlString());
-
-            return request;
-        }
-
-        public static void SalvarXmlEmDisco(this MDFeConsSitMDFe consSitMdfe)
-        {
-            if (MDFeConfiguracao.NaoSalvarXml()) return;
-
-            var caminhoXml = MDFeConfiguracao.CaminhoSalvarXml;
-
-            var arquivoSalvar = caminhoXml + @"\" + consSitMdfe.ChMDFe + "-ped-sit.xml";
-
-            FuncoesXml.ClasseParaArquivoXml(consSitMdfe, arquivoSalvar);
-        }
+        [XmlEnum("3.00")]
+        Versao300 = 300
     }
 }
