+++ MDFe.Utils/MDFe.Utils.csproj
@@ -36,38 +36,18 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Configuracoes\MDFeConfiguracao.cs" />
-    <Compile Include="Extencoes\ExtMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeConsReciMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeConsSitMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeConsStatServMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeCosMDFeNaoEnc.cs" />
-    <Compile Include="Extencoes\ExtMDFeEnviMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeEvCancMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeEvEncMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeEventoMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeEvIncCondutorMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetConsMDFeNao.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetConsReciMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetConsSitMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetConsStatServ.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetEnviMDFe.cs" />
-    <Compile Include="Extencoes\ExtMDFeRetEventoMDFe.cs" />
-    <Compile Include="Extencoes\ExtVersaoServico.cs" />
+    <Compile Include="Flags\VersaoServico.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Validacao\Validador.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DFe.Classes\DFe.Classes.csproj">
-      <Project>{9984EC15-18E1-4A87-B4DB-18AC913DC4E8}</Project>
+      <Project>{9984ec15-18e1-4a87-b4db-18ac913dc4e8}</Project>
       <Name>DFe.Classes</Name>
     </ProjectReference>
     <ProjectReference Include="..\DFe.Utils\DFe.Utils.csproj">
       <Project>{b4e1a7a6-5d66-4326-8a8d-39338e6fc84c}</Project>
       <Name>DFe.Utils</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\MDFe.Classes\MDFe.Classes.csproj">
-      <Project>{9794cb98-4691-4218-a127-c573c2e7ff83}</Project>
-      <Name>MDFe.Classes</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup />
